// src/app/models/pasien-register.model.ts
export class PasienRegister {
    id? : string;
    no_antrian: string;
    antrian: number;
    nama: string;
    alamat: string;
    no_telepon: string;
    nik: string;
    no_bpjs: string;
    poli: string;
    dokter: string;
    waktu: string;
    keluhan: string;
    status: string;
    createdAt: any;
}